fix href and add nofollow
